contributors/adobe/incoming/css3-regions/interactivity/mouse/mouse-events-region-alt-text.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>Mouse event: onmouseover</title>
        <link href="/resources/testharness.css" rel="stylesheet" type="text/css">
        <script src="/resources/testharness.js"></script>
        <style>
            /* Set the margin to 0 so we'll know that the target will be there if the content isn't flowing into a region */
            body {
                margin: 0;
            }
            /* Set the width of the content div so that the blocks will wrap downward when they're not flowing into a region */
            #content { 
              -webkit-flow-into: article;
              width: 75px; }
            /* Set regions to float left so the blocks will be laid out horizontally when they're flowing into the regions */
            .region{
              -webkit-flow-from: article;
              width: 75px;
              height: 75px;
              float:left;
            }
            /* Basic block */
            .block {
                background-color: gray;
                display: inline-block;
                width: 50px;
                height: 50px; 
            }
            /* Set color on the target to be different than the others */
            #target-block {
              background-color: rgb(0, 0, 0);
            }
            #log {
              clear: both;
            }
            /* This is arguably not really necessary since we're not verifying this anywhere */
            #target-block:hover { background-color: rgb(0, 255, 0); }
        </style>
    </head>
    <body>
        <p>Move the mouse over the black square then away from it.</p>
        <p>The test passes if all of the results below are PASS.</p>
        <div id="content">
            <div class="block" title="FAIL"></div>
            <div class="block" title="FAIL"></div>
            <div id="target-block" class="block" title="PASS"></div>
        </div>
        <div class="region"></div>
        <div class="region"></div>
        <div class="region"></div>
        <br/>
        <div id="log"></div>
    </body>
     <script type="text/javascript">

        function checkLeftPosition(elemID) 
        {
            return document.getElementById(elemID).getBoundingClientRect().left;
        }
        function checkBackgroundColor(elemID) {
            var foo = window.getComputedStyle(document.getElementById(elemID)).backgroundColor;
            return window.getComputedStyle(document.getElementById(elemID)).backgroundColor;
        }
        
        
        test(function() {assert_not_equals(checkLeftPosition("target-block"), 0)}, "Target is flowing into the region");

        var testTimeout = 15000;
        setup({timeout: testTimeout});

        var event_test1 = async_test("onmouseover event fired", {timeout: testTimeout});
            event_test1.step( function(){
                var testTarget = document.getElementById("target-block");
                testTarget.onmouseover = event_test1.step_func( function(evt) 
                {
                    /* Log assertion the event got fired */
                    assert_true(true);
                    
                    /* Verify the target was the one with the alt text set */
                    test(function() { assert_equals(evt.target.title, "PASS")}, "Target has the correct alt text");
                    
                    /* Verify the hover color is correct */
                    test(function() { assert_equals(checkBackgroundColor("target-block"), "rgb(0, 255, 0)") }, "Color changes on mouseover");
                    
                    event_test1.done();
                });
            });
         var event_test2 = async_test("onmouseout event fired", {timeout: testTimeout});
            event_test2.step( function(){
                var testTarget = document.getElementById("target-block");
                testTarget.onmouseout = event_test2.step_func( function(evt) 
                {
                    /* Log assertion the event got fired */
                    assert_true(true);

                    /* Verify it returns to the original background color */
                    test(function() { assert_equals(checkBackgroundColor("target-block"), "rgb(0, 0, 0)") }, "Color changes back on mouseout");

                    event_test2.done();
                });
            });
    </script>
</html>
